autoGen/requirement.txt
autogen-agentchat
autogen-ext[openai]
python-dotenv
asyncio
google-auth-oauthlib
google-api-python-client
PIL